Add documentation to all added/modified scripts
