+++ Assets/Scripts/Abstract/AbstractBillboard.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b81675e9956fea4498ac6a93a853547f
+guid: 0ba2e84f9fb52454c8de87089aaf9735
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
