+++ Assets/Scripts/Utils/NetworkUtils.cs
@@ -1,9 +1,16 @@
 using Mirror;
 using UnityEngine;
 
+/// <summary>
+/// A static utility class containing helper methods for network-related tasks.
+/// </summary>
 public static class NetworkUtils
 {
-    // TODO: Document
+    /// <summary>
+    /// Retrieves the <see cref="GameObject"/> representing the local player in the scene.
+    /// This is done by searching through all player controllers within the current scene and returns the one with a <see cref="NetworkIdentity"/> that is marked as the local player.
+    /// </summary>
+    /// <returns>The <see cref="GameObject"/> of the local player, if found; otherwise returns <c>null</c>.</returns>
     public static GameObject RetrieveLocalPlayer()
     {
         PlayerController[] playerControllers = GameObject.FindObjectsOfType<PlayerController>();
